backend/requirements.txt
# Python >= 3.10
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.0

# LangChain 0.3.x 全面支持 pydantic v2
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7

# pymilvus 2.5.x 移除了对 pydantic v1 的强依赖
pymilvus==2.5.0

openai==1.54.0
httpx==0.27.2
tiktoken==0.8.0
pypdf==5.1.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
# 认证相关
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.35
pymysql==1.1.1
email-validator==2.2.0
